More refactoring classes;
