+++ src/Base/Body.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ByJG\ApiTools;
+namespace ByJG\ApiTools\Base;
 
 use ByJG\ApiTools\Exception\GenericSwaggerException;
 use ByJG\ApiTools\Exception\InvalidRequestException;
@@ -8,7 +8,7 @@ use ByJG\ApiTools\Exception\NotMatchedException;
 use ByJG\ApiTools\Base\Schema;
 use InvalidArgumentException;
 
-abstract class SwaggerBody
+abstract class Body
 {
     const SWAGGER_PROPERTIES="properties";
     const SWAGGER_REQUIRED="required";
